server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==2.0.2
tensorflow-cpu==2.16.1
scikit-learn==1.6.1
joblib==1.5.2
